added helpers.js file for summary view; renumbered views
