+++ js/views/00_welcome/helpers.js
@@ -3,8 +3,8 @@ import Welcome from '../../components/welcome/welcome.js'
 import Body from '../../components/body/body.js'
 import Sidebar from '../../components/sidebar/sidebar.js'
 
-import { loadNewInitiatives } from '../06_new_initiatives/main.js'
-import { loadSummaryPage } from '../07_summary/main.js'
+import { loadNewInitiatives } from '../07_new_initiatives/main.js'
+import { loadSummaryPage } from '../08_summary/main.js'
 import { loadBaselineLandingPage } from '../02_baseline_landing_page/main.js'
 import { loadUploadPage } from '../01_upload/main.js'
