+++ js/views/07_new_initiatives/helpers.js
