+++ js/views/07_new_initiatives/main.js
