+++ js/views/08_summary/helpers.js
@@ -1,21 +1,10 @@
-import { CurrentPage } from "../../utils/data_utils/local_storage_handlers.js";
+
 import Prompt from '../../components/prompt/prompt.js'
 import Body from "../../components/body/body.js";
 import Subtitle from "../../components/header/header.js";
 import { visitPage } from "../view_logic.js";
 import { Accordion } from "../../components/accordion/accordion.js";
 import { downloadXLSX } from "../../utils/data_utils/XLSX_handlers.js";
-
-export function loadSummaryPage(){
-    //update page state
-    CurrentPage.update('summary');
-    summaryView();
-}
-
-export function cleanUpSummaryPage(){
-    Prompt.Buttons.Right.removeAction(returnToWelcome);
-    Prompt.Buttons.Left.removeAction(downloadXLSX);
-}
 
 export function summaryView(){
     Body.reset();
