+++ js/views/view_logic.js
@@ -1,11 +1,11 @@
 import { initializeWelcomePage } from './00_welcome/main.js';
-import { cleanUpInitiativesPage, loadNewInitiatives } from './06_new_initiatives/main.js'
+import { cleanUpInitiativesPage, loadNewInitiatives } from './07_new_initiatives/main.js'
 import { loadRevenuePage, cleanupRevenuePage } from './03_revenue/main.js'
 import { loadPersonnelPage } from './04_personnel/main.js';
-import { loadOTPage } from './04.5_OT/main.js';
+import { loadOTPage } from './05_overtime/main.js';
 import { loadNonpersonnelPage } from './05_nonpersonnel/main.js';
 import { loadBaselineLandingPage } from './02_baseline_landing_page/main.js';
-import { cleanUpSummaryPage, loadSummaryPage } from './07_summary/main.js';
+import { cleanUpSummaryPage, loadSummaryPage } from './08_summary/main.js';
 import { loadUploadPage } from './01_upload/main.js';
 import { pauseExecution } from '../utils/common_utils.js';
